+++ lib/flutter_bluetooth.dart
@@ -4,7 +4,7 @@ import 'dart:typed_data';
 import 'package:flutter/services.dart';
 import 'package:flutter_bluetooth/device_attributes.dart';
 import 'package:flutter_bluetooth/pair_device_response.dart';
-import 'package:flutter_bluetooth/paired_attributes.dart';
+import 'package:flutter_bluetooth/paired_device.dart';
 import 'package:flutter_bluetooth/socket_data.dart';
 
 import 'enable_bluetooth_response.dart';
@@ -25,23 +25,22 @@ class FlutterBluetooth {
   static Stream<Uint8List> read(String socketUUID) =>
       _eventChannel.receiveBroadcastStream("$socketUUID|1");
 
-  static Future<void> write(SocketData socketData) =>
+  static Future<bool> write(SocketData socketData) =>
       _channel.invokeMethod(OutgoingMethod.writeBt, socketData.toMap());
 
-  static Future<void> closeSocket(SocketData socketData) =>
+  static Future<bool> closeSocket(SocketData socketData) =>
       _channel.invokeMethod(OutgoingMethod.closeSocket, socketData.toMap());
 
   static Future<void> disconnect() =>
       _channel.invokeMethod(OutgoingMethod.disconnect);
 
-  static Future<PairedDeviceAttributes> connect(
-      DeviceAttributes attributes) async {
-    var completer = new Completer<PairedDeviceAttributes>();
+  static Future<PairedDevice> connect(DeviceAttributes attributes) async {
+    var completer = new Completer<PairedDevice>();
 
     _channel.setMethodCallHandler((call) {
       switch (call.method) {
         case PairDeviceResponse.devicePaired:
-          completer.complete(PairedDeviceAttributes.fromMap(call.arguments));
+          completer.complete(PairedDevice.fromMap(call.arguments));
           break;
         case PairDeviceResponse.cannotFindDevice:
           completer.completeError("Cannot find device");
@@ -60,7 +59,7 @@ class FlutterBluetooth {
     return completer.future;
   }
 
-  static Future<void> enable(Function() successCallback) async {
+  static Future<void> enable() async {
     var completer = new Completer();
 
     _channel.setMethodCallHandler((call) {