+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- improve write socket failure notification
+- improve close socket failure notification
+- added readme docs
+
 ## 0.0.1
 
 - init